+++ cachish/backends/__init__.py
@@ -1,1 +1,2 @@
 from .heroku import Heroku
+from .json_http import JsonHttp