+++ cachish/backends/heroku.py
@@ -1,7 +1,9 @@
 from requests import Session
 
+from .json_http import JsonHttp
 
-class Heroku(object):
+
+class Heroku(JsonHttp):
     def __init__(self,
             api_token=None,
             app=None,
@@ -10,18 +12,10 @@ class Heroku(object):
         assert api_token is not None
         assert app is not None
         assert config_key is not None
-        self.config_keys = [config_key] if isinstance(config_key, str) else config_key
-        self.session = Session()
-        self.session.trust_env = False # ignore .netrc auth
-        self.target_url = 'https://api.heroku.com/apps/%s/config-vars' % app
+
+        target_url = 'https://api.heroku.com/apps/%s/config-vars' % app
+        super(Heroku, self).__init__(url=target_url, field=config_key)
         self.session.headers.update({
             'Authorization': 'Bearer %s' % api_token,
             'Accept': 'application/vnd.heroku+json; version=3',
         })
-
-
-    def get(self):
-        response = self.session.get(self.target_url, timeout=5)
-        response.raise_for_status()
-        config = response.json()
-        return {key: config[key] for key in self.config_keys}