+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 
+UNRELEASED -
+------------------
+
+### Added
+- A generic JSON backend: JsonHttp. Configure with a url and field (possibly several) to extract.
+
+
 1.1.1 - 2017-01-19
 ------------------